requirements.txt
click==7.1.2
MouseInfo==0.1.3
numpy==1.20.3
opencv-python==4.5.2.52
Pillow==8.2.0
PyAutoGUI==0.9.52
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.8.2
pypiwin32==223
PyQt5==5.15.4
pyqt5-plugins==5.15.4.2.1.0
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
PyQt5-stubs==5.15.2.0
pyqt5-tools==5.15.4.3.0.3
PyRect==0.1.4
PyScreeze==0.1.27
python-dotenv==0.17.1
PyTweening==1.0.3
pywin32==300
qt5-applications==5.15.2.2.1
qt5-tools==5.15.2.1.0.1
